WATER-3409: Add change links on Charge information review page
